+++ FITI/FITI/Network/API/MyPageAPI.swift
@@ -1,0 +1,8 @@
+//
+//  MyPageAPI.swift
+//  FITI
+//
+//  Created by 홍준혁 on 2023/01/29.
+//
+
+import Foundation
