+++ FITI/FITI/Network/DTO/MyPage/HeartListResponse.swift
