+++ FITI/FITI/Network/DTO/MyPage/PetchLocationResponse.swift
@@ -5,7 +5,10 @@
 //  Created by 홍준혁 on 2023/01/26.
 //
 
-struct petchLocationModel: Codable {
+import Foundation
+
+// MARK: - Welcome
+struct PetchLocationResponse: Codable {
     let isSuccess: Bool
     let code: Int
     let message, result: String
