+++ FITI/FITI/Network/DTO/Trainer/TrainerArrayListRequest.swift
