+++ FITI/FITI/Network/DTO/Trainer/TrainerListResponse.swift
